+++ src/vgr_svg.ml
@@ -41,7 +41,7 @@ let save_gstate s = Set { g_tr = s.gstate.g_tr }
 let set_gstate s g = s.gstate <- g
 let view_rect s =          (* image view rect in current coordinate system. *)
   let tr = M3.inv s.gstate.g_tr in
-  Vgr.Private.Data.of_path (P.empty >> P.rect (Box2.tr tr s.view))
+  Vgr.Private.Data.of_path (P.empty |> P.rect (Box2.tr tr s.view))
 
 let image i = Vgr.Private.I.of_data i
 let partial = Vgr.Private.partial