+++ test/examples.ml
@@ -21,7 +21,7 @@ let svg_of_usquare i =
 let () = svg_of_usquare gray
 
 
-let circle = P.empty >> P.circle (P2.v 0.5 0.5) 0.4
+let circle = P.empty |> P.circle (P2.v 0.5 0.5) 0.4
 let gray_circle = I.cut circle gray
 
 let circle_outline =
@@ -30,31 +30,31 @@ let circle_outline =
   I.cut ~area circle black
 
 let dot = I.blend circle_outline gray_circle
-let dot = gray_circle >> I.blend circle_outline
+let dot = gray_circle |> I.blend circle_outline
 
 let scatter_plot pts pt_width =
   let dot =
-    let circle = P.empty >> P.circle P2.o (0.5 *. pt_width) in
-    I.const Color.black >> I.cut circle
+    let circle = P.empty |> P.circle P2.o (0.5 *. pt_width) in
+    I.const Color.black |> I.cut circle
   in
-  let mark pt = dot >> I.move pt in
-  let blend_mark acc pt = acc >> I.blend (mark pt) in
+  let mark pt = dot |> I.move pt in
+  let blend_mark acc pt = acc |> I.blend (mark pt) in
   List.fold_left blend_mark I.void pts
 
 let () =
   let p =
     let rel = true in
-    P.empty >>
-    P.sub (P2.v 0.1 0.5) >>
-    P.line (P2.v 0.3 0.5) >>
-    P.qcurve ~rel (P2.v 0.2 0.5) (P2.v 0.2 0.0) >>
-    P.ccurve ~rel (P2.v 0.0 (-. 0.5)) (P2.v 0.1 (-. 0.5)) (P2.v 0.3 0.0) >>
-    P.earc ~rel (Size2.v 0.1 0.2) (P2.v 0.15 0.0) >>
-    P.sub (P2.v 0.18 0.26) >>
-    P.qcurve ~rel (P2.v (0.01) (-0.1)) (P2.v 0.1 (-. 0.05)) >>
-    P.close >>
-    P.sub (P2.v 0.65 0.8) >>
+    P.empty |>
+    P.sub (P2.v 0.1 0.5) |>
+    P.line (P2.v 0.3 0.5) |>
+    P.qcurve ~rel (P2.v 0.2 0.5) (P2.v 0.2 0.0) |>
+    P.ccurve ~rel (P2.v 0.0 (-. 0.5)) (P2.v 0.1 (-. 0.5)) (P2.v 0.3 0.0) |>
+    P.earc ~rel (Size2.v 0.1 0.2) (P2.v 0.15 0.0) |>
+    P.sub (P2.v 0.18 0.26) |>
+    P.qcurve ~rel (P2.v (0.01) (-0.1)) (P2.v 0.1 (-. 0.05)) |>
+    P.close |>
+    P.sub (P2.v 0.65 0.8) |>
     P.line ~rel (P2.v 0.1 (-. 0.05))
   in
   let area = `O { P.o with P.width = 0.01 } in
-  ignore (I.const Color.black >> I.cut ~area p)
+  ignore (I.const Color.black |> I.cut ~area p)